refactor(runner.selenium): rename to gui from selenium
Looking forward to the potential of multiple types of runners in the
future, rename the package from "selenium" to "gui". This will more
clearly distinguish between future runner types (e.g. "cli" or "sdk")
and avoid coupling us to Selenium (in case we have a different GUI
automation/interaction technology in the future).
