+++ src/main/java/edu/iastate/ece/sd/sdmay2126/runner/gui/KBaseHelper.java
@@ -1,4 +1,4 @@
-package edu.iastate.ece.sd.sdmay2126.runner.selenium;
+package edu.iastate.ece.sd.sdmay2126.runner.gui;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
