+++ src/main/java/edu/iastate/ece/sd/sdmay2126/runner/gui/authentication/SeleniumAuthenticationFlows.java
@@ -1,4 +1,4 @@
-package edu.iastate.ece.sd.sdmay2126.runner.selenium.authentication;
+package edu.iastate.ece.sd.sdmay2126.runner.gui.authentication;
 
 /**
  * Indicates a particular 3rd-party authentication flow for KBase.
