+++ src/main/java/edu/iastate/ece/sd/sdmay2126/runner/gui/authentication/globus/GlobusAuthenticationFlow.java
@@ -1,8 +1,8 @@
-package edu.iastate.ece.sd.sdmay2126.runner.selenium.authentication.globus;
+package edu.iastate.ece.sd.sdmay2126.runner.gui.authentication.globus;
 
-import edu.iastate.ece.sd.sdmay2126.runner.selenium.SeleniumIdentificationException;
-import edu.iastate.ece.sd.sdmay2126.runner.selenium.SeleniumUtilities;
-import edu.iastate.ece.sd.sdmay2126.runner.selenium.authentication.SeleniumAuthenticationFlow;
+import edu.iastate.ece.sd.sdmay2126.runner.gui.SeleniumIdentificationException;
+import edu.iastate.ece.sd.sdmay2126.runner.gui.SeleniumUtilities;
+import edu.iastate.ece.sd.sdmay2126.runner.gui.authentication.SeleniumAuthenticationFlow;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
